# openapi.yaml
openapi: 3.1.0
info:
  title: CarLS — Carlsberg Learning & Support API
  version: 1.0.0
  description: |
    Опис усіх Route Handlers з `app/api/**/route.*` (2026-10-04).

    **Правило:** новий або змінений ендпоінт = правка цього файлу в тому ж PR.
    `lib/openapi.test.ts` валить CI, якщо роут і цей файл розходяться
    (шлях або метод є лише з одного боку), а `oasdiff` у PR позначає
    ламаючі зміни описаних контрактів.

    Повні схеми тіл — для зовнішніх контрактів (Power Query `/api/data/*`,
    Telegram, cron, health) і для відповіді на питання / завершення модуля.
    Решта (адмінка, кабінет) — шлях, метод, авторизація й коди відповідей:
    їхні клієнти живуть у цьому ж репозиторії й перевіряються рев'ю коду.

    Перегляд — `/admin/api` (Swagger UI, лише для супер-адміна).
servers:
  - url: /
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: session
      description: Сесія співробітника після входу кодом + PIN (lib/session.js). Підписана, зі sessionVersion — адмін може відкликати.
    adminSession:
      type: apiKey
      in: cookie
      name: admin_session
      description: Окремий вхід в /admin по ADMIN_PASSWORD (lib/adminSession.js), не пов'язаний із співробітником.
    superAdminSession:
      type: apiKey
      in: cookie
      name: admin_session
      description: admin_session рівня "super" (SUPER_ADMIN_PASSWORD).
    apiToken:
      type: http
      scheme: bearer
      description: Токен Power Query, виданий в /admin (AdminApiToken). Власник має бути активним admin/hr_manager.
    cronSecret:
      type: http
      scheme: bearer
      description: CRON_SECRET — Vercel Cron підставляє його сам.
    telegramSecret:
      type: apiKey
      in: header
      name: X-Telegram-Bot-Api-Secret-Token
      description: TELEGRAM_WEBHOOK_SECRET, заданий при реєстрації webhook.
  parameters:
    slug:
      name: slug
      in: path
      required: true
      schema: { type: string }
      description: Slug курсу
...
  responses:
    BadRequest:
      description: Невалідне тіло або параметри
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: Немає валідної сесії / ключа
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: Сесія є, але роль не дозволяє
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Не знайдено (або не належить цій людині)
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Conflict:
      description: Конфлікт стану
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Ok:
      description: Успіх (JSON)
      content: { application/json: { schema: { type: object } } }
    Created:
      description: Створено
      content: { application/json: { schema: { type: object } } }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    CourseState:
      type: object
      properties:
        completed: { type: boolean }
        scoreRaw: { type: integer }
        scoreMax: { type: integer }
        scorePercent: { type: integer }
        passed: { type: boolean }
        pointsEarned: { type: integer, description: "Бали за це складання (0 — курс уже був складений раніше або не складено)." }
      required: [completed, scoreRaw, scoreMax, scorePercent, passed, pointsEarned]
...
    TelegramUpdate:
      type: object
      description: "Апдейт Bot API (https://core.telegram.org/bots/api#update). Обробляються лише текстові повідомлення в приватному чаті."
      properties:
        update_id: { type: integer }
        message:
          type: object
          properties:
            text: { type: string, description: "/start <token> — прив'язати чат, /stop — відв'язати, інше — у журнал вхідних." }
            chat:
              type: object
              properties:
                id: { type: integer }
                type: { type: string }
            from:
              type: object
              properties:
                username: { type: string }
                first_name: { type: string }
paths:
...
  /api/courses/{slug}/submit:
    post:
      tags: [courses]
      deprecated: true
      summary: Закрити курс (лише для старої офлайн-черги — тепер курс закриває module-complete)
      security: [{ session: [] }]
      parameters: [{ $ref: "#/components/parameters/slug" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { enrollmentId: { type: integer } }
              required: [enrollmentId]
      responses:
        "200":
          description: Підсумок курсу
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  course: { $ref: "#/components/schemas/CourseState" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /api/admin/enrollments/{enrollmentId}:
    parameters: [{ name: enrollmentId, in: path, required: true, schema: { type: integer } }]
    patch:
      tags: [admin]
      summary: Ручна корекція проходження (статус, бал, дати; adminNote обов'язковий)
      security: [{ adminSession: [] }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [admin]
      summary: Зняти одне призначення (разом із балами)
      security: [{ adminSession: [] }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/admin/rating:
    get:
      tags: [admin]
      summary: Правила рейтингу (ваги) і рівні
      security: [{ adminSession: [] }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "403": { $ref: "#/components/responses/Forbidden" }
    put:
      tags: [admin]
      summary: Змінити правила рейтингу й рівні
      security: [{ adminSession: [] }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/admin/tokens:
    get:
      tags: [admin]
      summary: Токени Power Query
      security: [{ adminSession: [] }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [admin]
      summary: Видати токен Power Query (показується один раз)
      security: [{ adminSession: [] }]
      responses:
        "201": { $ref: "#/components/responses/Created" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/telegram/webhook:
    post:
      tags: [telegram]
      summary: Webhook бота — /start (прив'язати чат), /stop, інше → журнал вхідних
      description: Завжди 200 (крім невірного секрету) — інакше Telegram повторює апдейт безкінечно.
      security: [{ telegramSecret: [] }]
      requestBody:
        content: { application/json: { schema: { $ref: "#/components/schemas/TelegramUpdate" } } }
      responses:
        "200":
          description: Прийнято
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  reason: { type: string, enum: [not_configured] }
        "401": { $ref: "#/components/responses/Unauthorized" }
...